rename preprocessing/ to worker/ since worker will handle other stuff too
