+++ src/worker/mungers.js
